requirements.txt
astroid==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
pylint==2.3.1
PyYAML==5.1.2
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
